+++ packages/sewing-kit-koa/src/tests/middleware.test.ts
@@ -39,6 +39,22 @@ describe('middleware', () => {
     expect(getAssets(context)).toHaveProperty('userAgent', userAgent);
   });
 
+  it('defaults the manifest path to `build/client/assets.json`', async () => {
+    const context = createMockContext();
+    await middleware()(context, () => Promise.resolve());
+    expect(getAssets(context)).toHaveProperty(
+      'manifestPath',
+      'build/client/assets.json',
+    );
+  });
+
+  it('accespts a custom value for the manifest path', async () => {
+    const context = createMockContext();
+    const manifestPath = 'path/to/manifest';
+    await middleware({manifestPath})(context, () => Promise.resolve());
+    expect(getAssets(context)).toHaveProperty('manifestPath', manifestPath);
+  });
+
   it('calls the next middleware', async () => {
     const next = jest.fn(() => Promise.resolve());
     await middleware()(createMockContext(), next);